+++ src/ast.rs
@@ -1,0 +1,25 @@
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+pub enum BinaryOp {
+    Add,
+    Sub,
+    Mul,
+    Div,
+    LessThan,
+    LessThanOrEqual,
+    Equal,
+    NotEqual,
+}
+
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub enum Expr {
+    BinaryExpr {
+        op: BinaryOp,
+        op_pos: usize,
+        左辺: Box<Expr>,
+        右辺: Box<Expr>,
+    },
+    Numeric {
+        val: u8,
+        pos: usize,
+    },
+}
