+++ src/token.rs
@@ -1,0 +1,23 @@
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub enum TokenPayload {
+    Num(u8),
+    Add,
+    Sub,
+    Mul,
+    Div,
+    開き丸括弧,
+    閉じ丸括弧,
+    Eof,
+    GreaterThan,
+    GreaterThanOrEqual,
+    LessThan,
+    LessThanOrEqual,
+    Equal,
+    NotEqual,
+}
+
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub struct Token {
+    pub payload: TokenPayload,
+    pub pos: usize,
+}
